Move the UI reducer to the right place
